requirements.txt
absl-py==2.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
astunparse==1.6.3
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.3
cycler==0.12.1
fastapi==0.128.0
filelock==3.20.3
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2026.1.0
gast==0.7.0
google-pasta==0.2.0
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
humanfriendly==10.0
idna==3.11
Jinja2==3.1.6
keras==3.13.1
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.10.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
namex==0.1.0
networkx==3.6.1
numpy==2.4.1
onnx==1.20.1
onnxruntime==1.23.2
opencv-python==4.13.0.90
opt_einsum==3.4.0
optree==0.18.0
packaging==26.0
pillow==12.1.0
protobuf==6.33.4
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-multipart==0.0.21
requests==2.32.5
rich==14.2.0
six==1.17.0
starlette==0.50.0
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.3.0
torch==2.10.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
Werkzeug==3.1.5
wrapt==2.0.1


torchvision
pandas
kaggle
